Implemented basic large-domain ORE for n=64,k=8

+++ src/hash.rs
@@ -4,12 +4,11 @@ use aes::cipher::{
     BlockEncrypt, NewBlockCipher,
     generic_array::GenericArray,
 };
-use std::convert::From;
 
 /* Generates a short hash of the input.
  * Does not modify the input.
  * (uses AES in the Random Oracle model) */
-pub fn hash(key: &[u8], input: &[u8]) -> u128 { // TODO: can we use a U3 type?
+pub fn hash(key: &[u8], input: &[u8]) -> u8 { // TODO: can we use a U3 type?
     let key_array = GenericArray::from_slice(key);
     let cipher = Aes128::new(&key_array);
     // Can we clone into GenericArray directly? Are we doing an extra copy?
@@ -18,7 +17,8 @@ pub fn hash(key: &[u8], input: &[u8]) -> u128 { // TODO: can we use a U3 type?
     let mut block = GenericArray::from_mut_slice(&mut output);
     cipher.encrypt_block(&mut block);
     // Returns only the LSB (useful for 1-bit indicator function)
-    return u128::from(output[0]) & 1u128;
+    //return u128::from(output[0]) & 1u128;
+    return output[0] & 1u8;
 }
 
 // TODO: Test internal modules internally?
